Add missing return statements for method chaining
Also add addSort and setSize to metricAggregation to help with method
chaining.

This plugin has an established pattern where all setter methods return
the updated object so that more methods may be called on them in
sequence, some of these methods are missing this part and because of
that it is difficult to layout queries nicely.