Fix an issue where `installDomainMappings` is run repeatedly
Our service doesn't have any searchable domain classes, so
indexMap in ElasticAdminService will always be empty. The current check
in ElasticAdminService to see if domain mappings need to be installed
is to check is indexMap is empty, meaning on our case it runs on every
database insertion. This adds a check to make sure that
installDomainMappings is only run once.